monitoring-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
redis==5.2.1
celery==5.4.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
pydantic-settings==2.6.1
httpx==0.28.1
python-multipart==0.0.20
